+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityTableModelTest.java
@@ -114,10 +114,10 @@ public final class SwingEntityTableModelTest extends AbstractEntityTableModelTes
 
 	@Test
 	void editable() {
-		testModel.editable().set(true);
+		testModel.editor().enabled().set(true);
 		assertTrue(testModel.isCellEditable(0, 0));
 		assertFalse(testModel.isCellEditable(0, testModel.columns().identifiers().indexOf(Detail.INT_DERIVED)));
-		testModel.editable().set(false);
+		testModel.editor().enabled().set(false);
 	}
 
 	@Test
@@ -125,7 +125,7 @@ public final class SwingEntityTableModelTest extends AbstractEntityTableModelTes
 		SwingEntityTableModel tableModel = createTableModel(Employee.TYPE, connectionProvider());
 		tableModel.items().refresh();
 		assertThrows(IllegalStateException.class, () -> tableModel.setValueAt("newname", 0, 1));
-		tableModel.editable().set(true);
+		tableModel.editor().enabled().set(true);
 		tableModel.setValueAt("newname", 0, 1);
 		Entity entity = tableModel.items().included().get(0);
 		assertEquals("newname", entity.get(Employee.NAME));